UPDATE
UPdate:

+++ cgi-stu/stu/del.c
@@ -42,7 +42,20 @@ int cgiMain()
 		return -1;
 	}
 
-
+	sprintf(sql, "delete from selectcourse where id = %d", atoi(id));
+	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
+	{
+		fprintf(cgiOut,"mysql_real_query fail:%s\n", mysql_error(db));
+		mysql_close(db);
+		return -1;
+	}
+	sprintf(sql, "delete from score where id = %d", atoi(id));
+	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
+	{
+		fprintf(cgiOut,"mysql_real_query fail:%s\n", mysql_error(db));
+		mysql_close(db);
+		return -1;
+	}
 	sprintf(sql, "delete from information where id = %d", atoi(id));
 	if ((ret = mysql_real_query(db, sql, strlen(sql) + 1)) != 0)
 	{
